Fruits Keras Performance Upgrade
Improved performance drastically by making all the VGG16 layers trainable and using an Adam optimizer with a learning rate of 0.0001. The accuracy went from 36% (using only a fully-connected trainable layer at the end of the NN) to 99% (all the layers were trainable). Also added a confusion matrix to the script.